Working

+++ abc/abc151-200/src/abc186/abc186F.cpp
@@ -32,7 +32,7 @@ int main() {
 	for (int ti : t) {
 		fw.add(ti, 1);
 	}
-	int idx = 0;
+	long unsigned int idx = 0;
 	for (int i = 0; i < mh[0]; i++) {
 		while ((idx < t.size()) && (i >= mh[t[idx]])) {
 			fw.add(t[idx++], -1);
